Extract Shared Segmented Control
